Add unified_diff()
difflib.unified_diff offers a handy way to review the differences of two configuration files although it has a shortcoming for networking use cases. It presents a diff when commands are in a different order which is mostly irrelevant to network device configs, outside of ACLs, but they typically have sequence numbers. This version ignores order.